+++ controls.extend/propgrid/SPropertyItemBase.cpp
@@ -129,7 +129,7 @@ BOOL SPropertyItemBase::InsertChild(IPropertyItem *pChild, IPropertyItem *pInser
 
 int SPropertyItemBase::ChildrenCount() const
 {
-    return m_childs.GetCount();
+    return (int)m_childs.GetCount();
 }
 
 BOOL SPropertyItemBase::RemoveChild(IPropertyItem *pChild)
